--- a/src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/X509Certificates/X500NameEncoder.cs
+++ b/src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/X509Certificates/X500NameEncoder.cs
@@ -163,7 +163,7 @@ namespace System.Security.Cryptography.X509Certificates
             return (c == ' ' || (c >= 0x09 && c <= 0x0D));
         }
 
-        private static void AppendOid(StringBuilder decodedName, string oidValue)
+        private static void AppendOid(ref ValueStringBuilder decodedName, string oidValue)
         {
             Oid oid = new Oid(oidValue);